:bug: fix -> a lot of spell fixes and 1.0.0-alpha.1.6.11 version:
- flaming axe now inflicts fire effect in good sequence
- you can't put a magic slime when an entity already have a magic slime as passenger now
- phase step can teleport through walls now
- pocket furnace now only works when it actually smelts something
- changed fire element color from red to dark_red
- fixed items not able to have imbuement enchants on forge (related to Flaming Weapon not working)
- arc is marked as a sensible spell now
- conjure armor is marked as a sensible spell now
- lightning pulse is marked as a sensible spell now
- added missing particles on Zombie apocalypse

+++ common/src/main/java/com/binaris/wizardry/content/item/FlamingAxeItem.java
@@ -20,7 +20,7 @@ public class FlamingAxeItem extends AxeItem {
     @Override
     public boolean hurtEnemy(@NotNull ItemStack stack, @NotNull LivingEntity target, @NotNull LivingEntity entity) {
         if (!MagicDamageSource.isEntityImmune(EBDamageSources.FIRE, target))
-            target.setSecondsOnFire(Spells.FLAMING_AXE.property(DefaultProperties.EFFECT_DURATION).intValue());
+            target.setSecondsOnFire(Spells.FLAMING_AXE.property(DefaultProperties.EFFECT_DURATION));
         return false;
     }
